Cut down on needless casting per http://www.iso-9899.info/wiki/Casting

+++ date-time/c-delta-time.c
@@ -20,7 +20,7 @@ int main()
     assert(gettimeofday(&after, NULL) != -1);
 
     delta_t = after.tv_sec - before.tv_sec;
-    delta_t += (float) (after.tv_usec - before.tv_usec) / USEC_IN_SEC;
+    delta_t += (after.tv_usec - before.tv_usec) / USEC_IN_SEC;
     fprintf(stderr, "delta %.3f\n", delta_t);
 
     return (0);
